requirements.txt
django
django.tinymce
fontawesomefree
django-cripsy-forms   
django-recaptcha   
pillow      